first unit test of js buildFigureSentenceHelper()
